% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operate.R
\name{operate}
\alias{operate}
\title{Evaluate f with x and y}
\usage{
operate(x, y, f)
}
\arguments{
\item{x}{first input}

\item{y}{second input}

\item{f}{function to be evaluated using inputs x and y}
}
\description{
Evaluate f with x and y
}
